+++ .storybook/config.js
@@ -1,13 +1,16 @@
 import { configure, addDecorator, addParameters } from '@storybook/react';
 import { create } from '@storybook/theming';
-import React from 'react'
+import { withA11y } from '@storybook/addon-a11y';
+import logo from "../src/assets/FDS-logo.svg"
 
 import '../src/components/style/index.css';
 
+addDecorator(withA11y)
+
 const theme = create({
-  brandTitle: 'fds-components',
+  brandTitle: 'Form Design System',
   brandUrl: 'cbinsights.github.io/form-design-system',
-  brandImage: '', // We can add an actual logo later,
+  brandImage: logo, // We can add an actual logo later,
 });
 
 addParameters({
@@ -17,13 +20,15 @@ addParameters({
   ],
   options: {
     theme,
+    storySort: (a, b) =>
+      a[1].kind === b[1].kind ? 0 : a[1].id.localeCompare(b[1].id, { numeric: true })
   },
 });
 
-// automatically import all files ending in *.stories.js
-const req = require.context('../src/components', true, /\.stories\.js$/);
-function loadStories() {
-  req.keys().forEach((filename) => req(filename));
-}
-
-configure(require.context('../src/components', true, /\.stories\.(js|mdx)$/), module);
+configure(
+  [
+    require.context('../src/components', true, /\.docs\.stories\.(js|mdx)$/),
+    require.context('../src/components', true, /\.stories\.(js|mdx)$/),
+  ],
+  module
+);