+++ src/components/interactive/IconButton.jsx
@@ -29,7 +29,7 @@ const IconButton = ({
         'fdsIconButton',
         'rounded--all',
         'border--focus--noTransition',
-        'transition-default',
+        'transition--default',
         `fdsIconButton--${theme}`,
         `fdsIconButton--${size}`,
         {
@@ -41,6 +41,7 @@ const IconButton = ({
         }
       )}
       disabled={disabled && Element === 'button'}
+      role="img"
     >
       <span className={isLoading ? 'fdsIconButton--hidden' : ''}>
         <Icon customSize={size === 's' ? 16 : 18} />