+++ src/components/popovers/Popover.jsx
@@ -6,7 +6,6 @@ import { CSSTransition } from 'react-transition-group';
 import FDS from '../../../lib/dictionary/js/styleConstants';
 import { isNotRefsEvent } from '../util/events';
 
-const noop = () => {};
 export const VALID_POSITIONS = ['auto', 'top', 'right', 'bottom', 'left'];
 export const VALID_ALIGNMENTS = ['start', 'end', 'center'];
 export const VALID_INTERACTION_MODES = ['hover', 'click', 'controlled'];
@@ -201,8 +200,8 @@ Popover.defaultProps = {
   position: 'auto',
   alignment: 'start',
   distance: 4,
-  onOpen: noop,
-  onClose: noop,
+  onOpen: () => {},
+  onClose: () => {},
 };
 
 Popover.propTypes = {
@@ -235,12 +234,12 @@ Popover.propTypes = {
   disablePortal: PropTypes.bool,
 
   /**
-   * Places the popover content on the given side of the trigger.
+   * Vertical position preference of popover content.
    * `top` for example, will place the popover content above the trigger.
    */
   position: PropTypes.oneOf(VALID_POSITIONS),
 
-  /** Controls alignment of popover content relative to trigger */
+  /** Horizontal alignment preference of popover content relative to trigger */
   alignment: PropTypes.oneOf(VALID_ALIGNMENTS),
 
   /** Offset distance from trigger. */